mv the run job list creation in the prepare section
